+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added: Browser utility `CConnoisseur.closeNotes`.
 - Added: Cookie Connoisseur now waits for minigames to load by default.
+- Added: `verbose` setting in `cookie-connoisseur.config.json`.
 
 ## 0.2.1 - 2021-11-27
 - Added: Browser utilities `CConnoisseur.ascend` and `CConnoisseur.reincarnate`.